results/rebuttal/run-20231111-034248_LLMatic-zero_shot_donnemartin-system-design-oop-finance_0_20-runs_log/finance/4/LLMatic/data.py
class Transaction:
	def __init__(self, user_id, amount, category, month, year):
		self.user_id = user_id
		self.amount = amount
		self.category = category
		self.month = month
		self.year = year

class Data:
	def __init__(self):
		self.transactions = {}
		self.budgets = {}
		self.alerts = {}
		self.notifications = {}

	def add_transaction(self, user_id, amount, category, month, year):
		if user_id not in self.transactions:
			self.transactions[user_id] = []
		self.transactions[user_id].append(Transaction(user_id, amount, category, month, year))
		if user_id in self.budgets:
			self.budgets[user_id].add_amount(amount)
		self.check_for_alerts(user_id, amount)

	def import_transactions(self, user_id, transactions):
		if user_id not in self.transactions:
			self.transactions[user_id] = []
		for transaction in transactions:
			self.transactions[user_id].append(Transaction(user_id, transaction['amount'], transaction['category'], transaction['month'], transaction['year']))
			if user_id in self.budgets:
				self.budgets[user_id].add_amount(transaction['amount'])
			self.check_for_alerts(user_id, transaction['amount'])

	def check_for_alerts(self, user_id, amount):
		if user_id not in self.alerts:
			self.alerts[user_id] = []
		if amount < -500:
			self.alerts[user_id].append('Large transaction alert: $' + str(amount))

	def get_alerts(self, user_id):
		return self.alerts.get(user_id, [])

	def add_notification(self, user_id, notification):
		if user_id not in self.notifications:
			self.notifications[user_id] = []
		self.notifications[user_id].append(notification)

	def get_notifications(self, user_id):
		return self.notifications.get(user_id, [])

	def generate_monthly_report(self, user_id, month, year):
		# Placeholder for the implementation
		return {}

	def provide_spending_habits(self, user_id):
		# Placeholder for the implementation
		return {}

	def compare_year_on_year(self, user_id, year1, year2):
		# Placeholder for the implementation
		return {}
